+++ lib/ruby-next.rb
@@ -34,5 +34,6 @@ module RubyNext
   end
 
   require_relative "ruby-next/core"
+  require_relative "ruby-next/core_ext" if RubyNext::Core.core_ext?
   require_relative "ruby-next/logging"
 end